# platform/contracts/openapi/v1/wego-api.yaml
openapi: 3.1.1
info:
  title: Wego API
  version: 1.0.0
  description: >-
    Version 1 of the product-neutral Wego HTTP contract. WEGO-001 adds the
    first authenticated operations (login, logout, current identity);
    WEGO-002 adds the first business-domain operations (diving offerings and
    staff-created bookings).
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Current client instance
...
paths:
...
  /api/v1/divers/bookings/{id}:
    get:
      operationId: getBooking
      summary: Read a single booking
      tags:
        - Divers
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CorrelationIdHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The booking.
          headers:
            X-Correlation-Id:
              $ref: "#/components/headers/CorrelationIdResponseHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          description: No booking with this id.
  /api/v1/divers/bookings/{id}/cancel:
    post:
      operationId: cancelBooking
      summary: Cancel a confirmed booking
      description: >-
        Terminal — cancel only succeeds from CONFIRMED. Requires a
        non-blank `reason`. Independent of payment status: a paid booking
        that gets cancelled is not automatically refunded — refund is a
        separate, separately permissioned action.
      tags:
        - Divers
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CorrelationIdHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelBookingRequest"
      responses:
        "200":
          description: The booking was cancelled.
          headers:
            X-Correlation-Id:
              $ref: "#/components/headers/CorrelationIdResponseHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "400":
          description: A blank or oversized `reason`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          description: No booking with this id.
        "409":
          description: The booking is already cancelled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingError"
...
  /api/v1/tours-operator/bookings:
    post:
      operationId: createToursOperatorBooking
      summary: Create a temporary public booking
      description: >-
        Creates a NEW booking and reserves one place per guest. The price is
        computed by the server. It is not payment
        confirmation and must never be presented as CONFIRMED.
      tags:
        - ToursOperator
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateToursOperatorBookingRequest"
      responses:
        "201":
          description: NEW booking created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToursOperatorBookingResponse"
        "400":
          $ref: "#/components/responses/ToursOperatorValidationResponse"
        "404":
          description: Slot not found.
        "409":
          description: Slot blocked or already in the past (slot_in_past), not enough places left for the party, or tour inactive.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToursOperatorErrorResponse"
        "422":
          description: >-
            Pricing does not fit the tour — price_option_required,
            unit_count_required, guests_exceed_units, units_exceed_guests,
            price_option_not_applicable or child_price_not_available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToursOperatorErrorResponse"
        "503":
          description: >-
            Manager pause (bookings_paused) or enquiry-only deployment
            (online_booking_unavailable). No booking or inventory hold is created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToursOperatorErrorResponse"
    get:
      operationId: listToursOperatorBookings
      summary: List bookings for staff
      tags:
        - ToursOperator
      security:
        - bearerAuth: []
      parameters:
        - name: tourId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ToursOperatorBookingStatus"
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Staff-visible bookings.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ToursOperatorBookingResponse"
        "400":
          $ref: "#/components/responses/ToursOperatorValidationResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
  /api/v1/tours-operator/bookings/lookup:
    post:
      operationId: lookupToursOperatorBooking
      summary: Recover a booking by reference and phone
      description: >-
        Public recovery endpoint. The reference and phone are knowledge
        factors, not a staff identity. They are accepted only in the JSON body
        so PII is not placed in URLs; the edge applies a dedicated per-IP rate
        limit to reduce enumeration attempts.
      tags:
        - ToursOperator
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LookupToursOperatorBookingRequest"
      responses:
        "200":
          description: Matching booking.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToursOperatorPublicBookingLookupResponse"
        "400":
          $ref: "#/components/responses/ToursOperatorValidationResponse"
        "404":
          description: No matching booking.
        "429":
          $ref: "#/components/responses/RateLimitedResponse"
  /api/v1/tours-operator/bookings/{id}:
    get:
      operationId: getToursOperatorBooking
      summary: Get a booking for staff
      tags:
        - ToursOperator
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Booking details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToursOperatorBookingResponse"
        "400":
          $ref: "#/components/responses/ToursOperatorValidationResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          description: Booking not found.
  /api/v1/tours-operator/bookings/{id}/history:
    get:
      operationId: getToursOperatorBookingHistory
      summary: Booking lifecycle history for staff
      description: >
        Append-only lifecycle events oldest first: who changed the booking,
        when, and why. actorEmail is null for system or customer actions.
      tags:
        - ToursOperator
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Booking history, oldest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ToursOperatorBookingHistoryEntry"
        "400":
          $ref: "#/components/responses/ToursOperatorValidationResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          description: Booking not found.
  /api/v1/tours-operator/bookings/{id}/cancel:
    post:
      operationId: cancelToursOperatorBooking
      summary: Cancel a booking
      tags:
        - ToursOperator
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelToursOperatorBookingRequest"
      responses:
        "200":
          description: Cancelled booking.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToursOperatorBookingResponse"
        "400":
          $ref: "#/components/responses/ToursOperatorValidationResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          description: Booking not found.
        "409":
          description: Booking is already cancelled or cannot be cancelled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToursOperatorErrorResponse"
  /api/v1/tours-operator/bookings/{id}/complete:
    post:
      operationId: completeToursOperatorBooking
      summary: Mark a confirmed booking completed
      tags:
        - ToursOperator
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Completed booking.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToursOperatorBookingResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          description: Booking not found.
        "409":
          description: Booking cannot be completed from its current state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToursOperatorErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque session token issued by POST /api/v1/identity/login.
  responses:
...
    UnauthenticatedResponse:
      description: Missing, invalid, or expired session token.
      headers:
        WWW-Authenticate:
          description: Always `Bearer` — the authentication scheme this API accepts.
          schema:
            type: string
            enum:
              - Bearer
    ForbiddenResponse:
      description: Authenticated, but missing the permission this operation requires.
    RateLimitedResponse:
      description: >-
        Too many login attempts — from this network address (an edge-level
        limit) or against this email (an application-level limit that closes
        the same gap for a caller spreading attempts across many source
        addresses). Both return this identical shape.
      headers:
        Retry-After:
          description: Seconds to wait before the next attempt may succeed.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/LoginError"
    ToursOperatorValidationResponse:
      description: Malformed or domain-invalid tours-operator request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
  parameters:
...
    CorrelationIdHeader:
      name: X-Correlation-Id
      in: header
      required: false
      description: >-
        A caller-supplied correlation id, which must be a valid UUID.
        Echoed back on the response's `X-Correlation-Id` header and
        threaded through this request's audit and outbox records. If
        omitted or not a valid UUID, a new id is generated instead —
        malformed input never fails the request and never breaks
        traceability, it only replaces the id.
      schema:
        type: string
        format: uuid
  headers:
    CorrelationIdResponseHeader:
      description: >-
        The correlation id that applies to this request — the caller's
        own `X-Correlation-Id`, or a newly generated one if none was
        supplied or the supplied value was rejected. The same id appears
        on this request's audit and outbox records.
      schema:
        type: string
        format: uuid
  schemas:
...
    BookingStatus:
      type: string
      enum:
        - CONFIRMED
        - CANCELLED
    PaymentStatus:
      type: string
      enum:
        - UNPAID
        - PAID
        - REFUNDED
...
    ToursOperatorBookingStatus:
      type: string
      enum:
        - NEW
        - CONFIRMED
        - COMPLETED
        - CANCELLED
        - EXPIRED
...
    CreateToursOperatorBookingRequest:
      type: object
      additionalProperties: false
      required:
        - slotId
        - adultsCount
        - childrenCount
        - customer
        - hotelName
        - locale
      properties:
        slotId:
          type: string
          format: uuid
        adultsCount:
          type: integer
          minimum: 1
          maximum: 50
        childrenCount:
          type: integer
          minimum: 0
          maximum: 50
        priceOptionCode:
          description: Per-unit tours only — the chosen price option.
          type: string
          maxLength: 40
          pattern: "^[a-z0-9][a-z0-9-]{0,39}$"
        unitCount:
          description: Per-unit tours only — how many units; guests must fit in them.
          type: integer
          minimum: 1
          maximum: 50
        customer:
          $ref: "#/components/schemas/ToursOperatorCustomerRequest"
        hotelName:
          type: string
          minLength: 1
          maxLength: 200
        hotelRoom:
          type: string
          maxLength: 32
        specialRequests:
          type: string
          maxLength: 4000
        locale:
          type: string
          enum:
            - en
            - ar
            - ru
            - it
    LookupToursOperatorBookingRequest:
      type: object
      additionalProperties: false
      required:
        - reference
        - phone
      properties:
        reference:
          type: string
          minLength: 1
          maxLength: 64
          pattern: "^[Ss][Tt][Rr]-[0-9]{4}-[0-9]+$"
        phone:
          type: string
          minLength: 1
          maxLength: 32
    ToursOperatorBookingResponse:
      type: object
      additionalProperties: false
      required:
        - id
        - reference
        - tourId
        - slotId
        - tourDate
        - timeSlot
        - adultsCount
        - childrenCount
        - priceAdult
        - priceChild
        - totalPrice
        - unit
        - customer
        - hotelName
        - hotelRoom
        - specialRequests
        - locale
        - status
        - createdAt
        - confirmedAt
        - cancelledAt
        - cancellationReason
        - completedAt
        - expiredAt
      properties:
        id:
          type: string
          format: uuid
        reference:
          type: string
          pattern: "^STR-[0-9]{4}-[0-9]+$"
        tourId:
          type: string
          format: uuid
        slotId:
          type: string
          format: uuid
        tourDate:
          type: string
          format: date
        timeSlot:
          $ref: "#/components/schemas/ToursOperatorTimeSlot"
        adultsCount:
          type: integer
          minimum: 1
        childrenCount:
          type: integer
          minimum: 0
        priceAdult:
          $ref: "#/components/schemas/Money"
        priceChild:
          anyOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
        totalPrice:
          $ref: "#/components/schemas/Money"
        unit:
          anyOf:
            - $ref: "#/components/schemas/ToursOperatorBookingUnit"
            - type: "null"
        customer:
          $ref: "#/components/schemas/ToursOperatorCustomerResponse"
        hotelName:
          type: string
        hotelRoom:
          type:
            - string
            - "null"
        specialRequests:
          type:
            - string
            - "null"
        locale:
          type: string
        status:
          $ref: "#/components/schemas/ToursOperatorBookingStatus"
        createdAt:
          type: string
          format: date-time
        confirmedAt:
          type:
            - string
            - "null"
          format: date-time
        cancelledAt:
          type:
            - string
            - "null"
          format: date-time
        cancellationReason:
          type:
            - string
            - "null"
        completedAt:
          type:
            - string
            - "null"
          format: date-time
        expiredAt:
          type:
            - string
            - "null"
          format: date-time
    ToursOperatorPublicBookingLookupResponse:
      type: object
      additionalProperties: false
      required:
        - reference
        - tourDate
        - timeSlot
        - adultsCount
        - childrenCount
        - totalPrice
        - unit
        - hotelName
        - status
        - cancellationReason
      properties:
        reference:
          type: string
          pattern: "^STR-[0-9]{4}-[0-9]+$"
        tourDate:
          type: string
          format: date
        timeSlot:
          $ref: "#/components/schemas/ToursOperatorTimeSlot"
        adultsCount:
          type: integer
          minimum: 1
        childrenCount:
          type: integer
          minimum: 0
        totalPrice:
          $ref: "#/components/schemas/Money"
        unit:
          anyOf:
            - $ref: "#/components/schemas/ToursOperatorBookingUnit"
            - type: "null"
        hotelName:
          type: string
        status:
          $ref: "#/components/schemas/ToursOperatorBookingStatus"
        cancellationReason:
          type:
            - string
            - "null"
    CancelToursOperatorBookingRequest:
      type: object
      additionalProperties: false
      required:
        - reason
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 1000
...
    ToursOperatorErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: string
...
    ToursOperatorBookingHistoryEntry:
      type: object
      additionalProperties: false
      required:
        - eventType
        - occurredAt
      properties:
        eventType:
          type: string
          enum:
            - BOOKING_CREATED
            - BOOKING_CONFIRMED
            - BOOKING_CANCELLED
            - BOOKING_COMPLETED
            - BOOKING_EXPIRED
        fromStatus:
          type:
            - string
            - "null"
        toStatus:
          type:
            - string
            - "null"
        reason:
          type:
            - string
            - "null"
        actorUserId:
          type:
            - string
            - "null"
          format: uuid
        actorEmail:
          type:
            - string
            - "null"
        occurredAt:
          type: string
          format: date-time
...
    BookingResponse:
      type: object
      additionalProperties: false
      required:
        - id
        - offeringId
        - partySize
        - customerName
        - status
        - paymentStatus
        - pricingBasis
        - unitPrice
        - billableQuantity
        - totalPrice
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        offeringId:
          type: string
          format: uuid
        partySize:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        status:
          $ref: "#/components/schemas/BookingStatus"
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        pricingBasis:
          $ref: "#/components/schemas/PricingBasis"
        unitPrice:
          allOf:
            - $ref: "#/components/schemas/Money"
          description: >-
            The per-unit price snapshotted from the offering at the time
            this booking was created — unaffected by any later change to
            the offering's own price.
        billableQuantity:
          type: integer
          description: >-
            `partySize` when `pricingBasis` is `PER_PARTICIPANT`, always
            `1` when `pricingBasis` is `FLAT`.
        totalPrice:
          allOf:
            - $ref: "#/components/schemas/Money"
          description: "`unitPrice.amount * billableQuantity`, snapshotted at creation time."
        createdAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
        cancellationReason:
          type: string
    CancelBookingRequest:
      type: object
      additionalProperties: false
      required:
        - reason
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 1000
...
    ValidationErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
        - message
      properties:
        error:
          type: string
          enum:
            - validation_failed
        message:
          type: string
          description: A safe, human-readable description of which rule failed.
    BookingError:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: string
          enum:
            - offering_closed
            - already_closed
            - capacity_exceeded
            - already_cancelled
            - invalid_payment_transition
            - idempotency_key_conflict
...